# apps/backend/api/swagger.yaml
basePath: /
definitions:
  storage.Avatar:
    properties:
      md:
        type: string
      raw:
        type: string
      sm:
        type: string
    type: object
  storage.CharacterWithRelationInfo:
    properties:
      _id:
        type: string
      aliases:
        type: string
      avatar:
        $ref: '#/definitions/storage.Avatar'
      bio:
        type: string
      class:
        type: string
      name:
        type: string
      race:
        type: string
      relation:
        properties:
          objective_type:
            type: string
          role:
            type: string
        type: object
    type: object
  storage.ChildQuest:
    properties:
      _id:
        type: string
      _key:
        type: string
      body:
        type: string
      images:
        $ref: '#/definitions/storage.QuestImages'
      title:
        type: string
    type: object
  storage.Note:
    properties:
      _id:
        type: string
      body:
        type: string
      relation:
        additionalProperties: true
        type: object
      title:
        type: string
    type: object
  storage.QuestDetail:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      body:
        type: string
      characters:
        items:
          $ref: '#/definitions/storage.CharacterWithRelationInfo'
        type: array
      children:
        items:
          $ref: '#/definitions/storage.ChildQuest'
        type: array
      images:
        $ref: '#/definitions/storage.QuestImages'
      notes:
        items:
          $ref: '#/definitions/storage.Note'
        type: array
      title:
        type: string
    type: object
  storage.QuestImages:
    properties:
      banner:
        type: string
    type: object
  storage.QuestWithLinkedCharacters:
    properties:
      _id:
        type: string
      _key:
        type: string
      body:
        type: string
      characters:
        items:
          properties:
            avatar:
              $ref: '#/definitions/storage.Avatar'
            name:
              type: string
            objective_type:
              type: string
            role:
              type: string
          type: object
        type: array
      images:
        $ref: '#/definitions/storage.QuestImages'
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: alex@alexmeuer.com
    name: Alex Meuer
  description: This relies on an ArangoDB database for most operations.
  title: Simply-the-Quest HTTP API
  version: "1.0"
paths:
  /ping:
    get:
      consumes:
      - text/plain; charset=utf-8
      produces:
      - text/plain; charset=utf-8
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Send Ping, get Pong
  /quests:
    get:
      consumes:
      - application/json
      description: Gets a page of quests with basic relations
      parameters:
      - default: 10
        description: limit
        format: int
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: offset
        format: int
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.QuestWithLinkedCharacters'
            type: array
      summary: List quests
      tags:
      - quests
  /quests/:id:
    get:
      consumes:
      - application/json
      description: Gets details of a quest
      parameters:
      - description: quest id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.QuestDetail'
            type: array
      summary: Quest details
      tags:
      - quests
swagger: "2.0"